refactor/#245: usePetProfileStep 안에 useNavigate 포함
